move retrieval conf
