+++ packages/chatbot-server-mongodb-public/src/config.ts
@@ -96,7 +96,8 @@ export const embeddedContentStore = makeMongoDbEmbeddedContentStore({
 });
 
 export const retrievalConfig = {
-  model: OPENAI_EMBEDDING_DEPLOYMENT,
+  preprocessorLlm: OPENAI_PREPROCESSOR_CHAT_COMPLETION_DEPLOYMENT,
+  embeddingModel: OPENAI_EMBEDDING_DEPLOYMENT,
   findNearestNeighborsOptions: {
     k: 5,
     path: "embedding",
@@ -107,7 +108,7 @@ export const retrievalConfig = {
 
 export const embedder = makeOpenAiEmbedder({
   openAiClient,
-  deployment: retrievalConfig.model,
+  deployment: retrievalConfig.embeddingModel,
   backoffOptions: {
     numOfAttempts: 3,
     maxDelay: 5000,
@@ -161,7 +162,7 @@ export const generateUserPrompt = wrapTraced(
     onNoVerifiedAnswerFound: wrapTraced(
       makeStepBackRagGenerateUserPrompt({
         openAiClient: preprocessorOpenAiClient,
-        model: OPENAI_PREPROCESSOR_CHAT_COMPLETION_DEPLOYMENT,
+        model: retrievalConfig.preprocessorLlm,
         findContent,
         numPrecedingMessagesToInclude: 6,
       }),
