+++ packages/chatbot-server-mongodb-public/src/eval/evalHelpers.ts
@@ -5,7 +5,7 @@ import {
   UserMessage,
 } from "mongodb-chatbot-server";
 import { EVAL_ENV_VARS } from "../EvalEnvVars";
-import { AzureOpenAI } from "openai";
+import { AzureOpenAI } from "mongodb-rag-core/openai";
 import { z } from "zod";
 import { strict as assert } from "assert";
